+++ src/main/kotlin/blackjack/BlackJackApplication.kt
@@ -1,5 +1,9 @@
 package blackjack
 
+import blackjack.domain.BlackJackGame
+import blackjack.view.InputView
+import blackjack.view.OutputView
+
 fun main() {
     val inputView = InputView()
     val outputView = OutputView()
