+++ src/main/kotlin/blackjack/domain/BlackJackGame.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.domain
 
 class BlackJackGame(
     val players: List<Player>,
