+++ src/test/kotlin/blackjack/domain/DeckTest.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.domain
 
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.types.shouldBeInstanceOf
